Replace automated script with manual install steps to fix data loss (#13)
@TolichP [1] reported in arcticicestudio/nord-terminator#3 [2
`install.sh` script replaced the `config` file which led to
configuration/data loss. This was caused by the `-f` flag passed to the
`cp` command [3] which forced to override existing files.

This is a really critical bug and the the solution to provide a script
that automatically handles the simple task of copying a file has never
been a good idea at all. Therefore the script was removed and the
documentation has been adjusted to provide manual installation steps
instead like in other port projects, e.g. Nord Alacritty [4].
The manual steps are quite simple and the knowledge can be expected of
users since these are basic operations when running a Unix system.

Projects like Terminato, that require to modify the main user
configuration file instead of providing a way to load features like
themes from dedicated file, are always way harder to support as a
theme author. Anyway, there are plans to unify the installation of Nord
port projects through one universal automation tool,
but there is currently no ETA.

[1]: https://github.com/TolichP
[2]: https://github.com/arcticicestudio/nord-terminator/issues/3#issuecomment-610723008
[3]: https://github.com/arcticicestudio/nord-terminator/blob/af29e150a531f7984d362f4aa135c1bfffd13c45/install.sh#L83
[4]: https://github.com/arcticicestudio/nord-alacritty

Fixes GH-12